+++ orc-rt/test/unit/CMakeLists.txt
@@ -7,11 +7,22 @@ if (NOT TARGET llvm_gtest)
   return ()
 endif ()
 
-function(add_orc_rt_unittest test_dirname)
-  add_unittest(OrcRTUnitTests ${test_dirname} ${ARGN})
+# Adds a unit test suite for one layer. link_lib is the layer's library; the
+# remaining arguments are sources.
+function(add_orc_rt_unittest test_name link_lib)
+  add_unittest(OrcRTUnitTests ${test_name} ${ARGN} DISABLE_LLVM_LINK_LLVM_DYLIB)
+  target_compile_options(${test_name} PRIVATE ${ORC_RT_COMPILE_FLAGS})
+  # Shared test helpers (CommonTestUtils.h, DirectCaller.h, ...) live at the
+  # root of this directory, so tests in the layer subdirectories can include
+  # them by bare name.
+  target_include_directories(${test_name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+  target_link_libraries(${test_name} PRIVATE ${link_lib})
 endfunction()
 
-add_orc_rt_unittest(CoreTests
+# SupportTests links the Support objects and nothing else, so a new
+# support-to-bedrock dependency shows up here as a link error rather than
+# going unnoticed.
+add_orc_rt_unittest(SupportTests orc-rt-support
   support/AllocActionTest.cpp
   support/BitmaskEnumTest.cpp
   support/CAPICompileTest.c
@@ -49,6 +60,10 @@ add_orc_rt_unittest(CoreTests
 
   support/sys/ErrnoTest.cpp
 
+  tools/OptionParserTest.cpp
+  )
+
+add_orc_rt_unittest(BedrockTests orc-rt-bedrock
   bedrock/BootstrapInfoTest.cpp
   bedrock/ExecutorProcessInfoTest.cpp
   bedrock/InProcessControllerAccessTest.cpp
@@ -69,18 +84,6 @@ add_orc_rt_unittest(CoreTests
 
   bedrock/sys/CPUFeaturesTest.cpp
   bedrock/sys/TargetTripleTest.cpp
-
-  tools/OptionParserTest.cpp
-
-  DISABLE_LLVM_LINK_LLVM_DYLIB
-  )
-target_compile_options(CoreTests PRIVATE ${ORC_RT_COMPILE_FLAGS})
-# Cross-layer test helpers (CommonTestUtils.h, DirectCaller.h, ...) live at the
-# root of this directory, so tests in the layer subdirectories can include them
-# by bare name.
-target_include_directories(CoreTests PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
-target_link_libraries(CoreTests PRIVATE
-  orc-rt-bedrock
   )
 
 # Build a shared library for NativeDylibManager tests.
@@ -88,6 +91,6 @@ add_library(NativeDylibManagerTestLib SHARED
   Inputs/NativeDylibManagerTestLib.cpp)
 set_target_properties(NativeDylibManagerTestLib PROPERTIES
   PREFIX "")
-target_compile_definitions(CoreTests PRIVATE
+target_compile_definitions(BedrockTests PRIVATE
   "NDM_TEST_LIB_PATH=\"$<TARGET_FILE:NativeDylibManagerTestLib>\"")
-add_dependencies(CoreTests NativeDylibManagerTestLib)
+add_dependencies(BedrockTests NativeDylibManagerTestLib)
