+++ orc-rt/test/unit/CommonTestUtils.h
@@ -9,8 +9,10 @@
 #ifndef ORC_RT_UNITTEST_COMMONTESTUTILS_H
 #define ORC_RT_UNITTEST_COMMONTESTUTILS_H
 
-#include "orc-rt/bedrock/ExecutorProcessInfo.h"
-#include "orc-rt/bedrock/Session.h"
+// Helpers here must not depend on Bedrock: this header is included by
+// SupportTests translation units, which link Support alone. Bedrock-dependent
+// helpers belong in BedrockTestUtils.h.
+
 #include "orc-rt/support/Error.h"
 #include "orc-rt/support/WrapperFunction.h"
 #include "orc-rt/support/move_only_function.h"
@@ -42,24 +44,6 @@ public:
 private:
   std::vector<std::string> &ErrMsgs;
 };
-
-inline orc_rt::ExecutorProcessInfo mockExecutorProcessInfo() noexcept {
-  return orc_rt::ExecutorProcessInfo("arm64-apple-darwin", 16384,
-                                     "+neon, +fullfp16");
-}
-
-/// DispatchFn for tests that should never dispatch a task. Records a test
-/// failure on invocation, then runs the task inline so that any caller
-/// awaiting a result unblocks (rather than hanging) and the keepalive token
-/// is released, even in -Asserts builds or when the dispatch arrives on a
-/// non-test thread.
-inline void noDispatch(orc_rt::Session::Task T) {
-  ADD_FAILURE() << "unexpected dispatch in a no-dispatch session";
-  T();
-}
-
-/// DispatchFn that runs tasks on the current thread.
-inline void inlineDispatch(orc_rt::Session::Task T) { T(); }
 
 template <size_t Idx = 0> class OpCounter {
 public:
