+++ orc-rt/test/unit/bedrock/NativeDylibManagerTest.cpp
@@ -15,6 +15,7 @@
 
 #include "gtest/gtest.h"
 
+#include "BedrockTestUtils.h"
 #include "CommonTestUtils.h"
 
 #include <optional>
