+++ orc-rt/test/unit/bedrock/sps/NativeDylibManagerSPSCITest.cpp
@@ -15,6 +15,7 @@
 #include "orc-rt/bedrock/Session.h"
 #include "orc-rt/support/sps/SPSWrapperFunction.h"
 
+#include "BedrockTestUtils.h"
 #include "CommonTestUtils.h"
 #include "DirectCaller.h"
 #include "gtest/gtest.h"
